Changes made in this release: - The BuildHelp and CleanHelp tasks now log all messages with an importance of High.  If set to Normal and ran from within Visual Studio, it won't show the progress messages when the logging options are set to Minimal. Note that this does not apply when built from within a pre-build or post-build task.  In those cases, the output is only shown once the commands complete. - The Microsoft.Common.targets file is now imported in the SandcastleHelpFileBuilder.targets file.  This is required as the Managed Package Framework expects access to some of the common targets.  It also allows COM references to be resolved when added in Visual Studio. - In order to prevent the common targets from setting the OutDir property to the SHFB OutputPath value, a default OutDir value is defined in the SandcastleHelpFileBuilder.targets file (.\).  SHFB will ignore this value if it sees it. - The BuildHelp MSBuild task is now cancelable.  Note that if canceled while BuildAssembler is initializing that it may take 10-20 seconds for MSBuild to stop. - Made various changes to the build engine to allow it to work when called from within Visual Studio. - Moved the IndentHtml property to the Build category since it is a better fit there. - Renamed the "Comments" property category "Summaries" to better match its purpose. - Added the SHFB Project Launcher that lets you choose how SHFB projects are opened. - Added the initial version of the VSPackage to provide Visual Studio integration for SHFB. - Added solution files for the documentation projects since they can be loaded in Visual Studio now. - Updated the SHFB help file topics with information on the extension package.

+++ Main/Source/SandcastleBuilderUtils/BuildEngine/BuildProcess.cs
@@ -2,7 +2,7 @@
 // System  : Sandcastle Help File Builder Utilities
 // File    : BuildProcess.cs
 // Author  : Eric Woodruff  (Eric@EWoodruff.us)
-// Updated : 03/06/2011
+// Updated : 04/09/2011
 // Note    : Copyright 2006-2011, Eric Woodruff, All rights reserved
 // Compiler: Microsoft Visual C#
 //
@@ -72,6 +72,7 @@ using System.Web;
 using System.Xml;
 using System.Xml.XPath;
 
+using SandcastleBuilder.Utils.BuildComponent;
 using SandcastleBuilder.Utils.ConceptualContent;
 using SandcastleBuilder.Utils.Design;
 using SandcastleBuilder.Utils.MSBuild;
@@ -154,6 +155,11 @@ namespace SandcastleBuilder.Utils.BuildEngine
 
         #region Properties
         //=====================================================================
+
+        /// <summary>
+        /// This is used to get the cancellation state of the build
+        /// </summary>
+        public bool BuildCanceled { get; set; }
 
         /// <summary>
         /// This returns the path to MSBuild.exe
@@ -557,8 +563,7 @@ namespace SandcastleBuilder.Utils.BuildEngine
                 Assembly asm = Assembly.GetExecutingAssembly();
 
                 FileVersionInfo fvi = FileVersionInfo.GetVersionInfo(asm.Location);
-                this.ReportProgress(BuildStep.Initializing,
-                    "[{0}, version {1}]", fvi.ProductName, fvi.ProductVersion);
+                this.ReportProgress(BuildStep.Initializing, "[{0}, version {1}]", fvi.ProductName, fvi.ProductVersion);
 
                 buildStart = stepStart = DateTime.Now;
 
@@ -566,7 +571,7 @@ namespace SandcastleBuilder.Utils.BuildEngine
                     t => t.ToolsVersion == project.MSBuildProject.ToolsVersion).ToolsPath, "MSBuild.exe");
 
                 // Base folder for SHFB
-                shfbFolder = Path.GetDirectoryName(asm.Location) + @"\";
+                shfbFolder = BuildComponentManager.HelpFileBuilderFolder;
 
                 // Get the location of the template files
                 templateFolder = shfbFolder + @"Templates\";
@@ -1429,7 +1434,7 @@ AllDone:
                     currentProcess.Kill();
 
                     foreach(Process p in Process.GetProcesses())
-                        if(reKillProcess.IsMatch(p.ProcessName))
+                        if(reKillProcess.IsMatch(p.ProcessName) && !p.HasExited)
                         {
                             System.Diagnostics.Debug.WriteLine("Killing " + p.ProcessName);
                             p.Kill();
@@ -1651,7 +1656,11 @@ AllDone:
 
             if(stepChanged)
                 OnBuildStepChanged(progressArgs);
+
+            if(this.BuildCanceled && !progressArgs.HasCompleted)
+                throw new BuilderException("BUILD CANCELLED");
         }
+
         #endregion
 
         #region Helper methods
@@ -2520,6 +2529,16 @@ AllDone:
             {
                 System.Diagnostics.Debug.WriteLine(ex);
                 System.Diagnostics.Debug.WriteLine("ReadStdOut thread failed\r\n");
+
+                try
+                {
+                    if(currentProcess != null)
+                        currentProcess.Kill();
+                }
+                catch
+                {
+                    // Ignore exceptions as the process may already have exited
+                }
             }
 
             System.Diagnostics.Debug.WriteLine("ReadStdOut thread stopped\r\n");
@@ -2551,6 +2570,16 @@ AllDone:
             {
                 System.Diagnostics.Debug.WriteLine(ex);
                 System.Diagnostics.Debug.WriteLine("ReadStdErr thread failed\r\n");
+
+                try
+                {
+                    if(currentProcess != null)
+                        currentProcess.Kill();
+                }
+                catch
+                {
+                    // Ignore exceptions as the process may already have exited
+                }
             }
 
             System.Diagnostics.Debug.WriteLine("ReadStdErr thread stopped\r\n");
